Applied the new Carousel on the old one
