+++ src/components/providers/CarouselProvider.tsx
@@ -1,444 +1,143 @@
-import BrokenImageOutlinedIcon from '@mui/icons-material/BrokenImageOutlined';
-import CloseOutlinedIcon from '@mui/icons-material/CloseOutlined';
-import NavigateBeforeIcon from '@mui/icons-material/NavigateBefore';
-import NavigateNextIcon from '@mui/icons-material/NavigateNext';
-import { alpha, Box, CircularProgress, Drawer, IconButton, Tooltip, Typography } from '@mui/material';
-import makeStyles from '@mui/styles/makeStyles';
-import clsx from 'clsx';
-import Carousel from 'commons/addons/carousel/Carousel';
-import useMyAPI from 'components/hooks/useMyAPI';
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import useALContext from 'components/hooks/useALContext';
+import { CarouselContainer } from 'components/visual/Carousel/Container';
+import SimpleSearchQuery from 'components/visual/SearchBar/simple-search-query';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { useLocation, useNavigate } from 'react-router';
 
-const XLSize = '96px';
-const LGSize = '80px';
-const MDSize = '64px';
-
-const useStyles = makeStyles(theme => ({
-  backdrop: {
-    backgroundColor: 'transparent',
-    backdropFilter: 'blur(1px)',
-    zIndex: 1350
-  },
-  carousel: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    flexWrap: 'nowrap',
-    height: '100%',
-    '@media (min-height:500px)': {
-      paddingTop: theme.spacing(2),
-      paddingBottom: theme.spacing(2)
-    }
-  },
-  spacer: {
-    flexGrow: 1
-  },
-  text: {
-    color: 'white',
-    display: 'none',
-    '@media (min-height:500px)': {
-      display: 'block'
-    },
-    textAlign: 'center'
-  },
-  textContainer: {
-    maxWidth: '90vh'
-  },
-  imageContainer: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    minHeight: theme.spacing(24),
-    minWidth: theme.spacing(24)
-  },
-  image: {
-    objectFit: 'contain',
-    minHeight: theme.spacing(12),
-    minWidth: theme.spacing(12),
-    maxWidth: '100vw',
-    maxHeight: `calc(95vh - 2*${theme.spacing(3)} - 2*${theme.spacing(2)})`,
-    '@media (min-height:500px)': {
-      maxHeight: `calc(95vh - ${MDSize} - 2*${theme.spacing(3)} - 2*${theme.spacing(2)})`
-    },
-    '@media (min-height:720px)': {
-      maxHeight: `calc(95vh - ${LGSize} - 2*${theme.spacing(3)} - 2*${theme.spacing(2)})`
-    },
-    '@media (min-height:1080px)': {
-      maxHeight: `calc(95vh - ${XLSize} - 2*${theme.spacing(3)} - 2*${theme.spacing(2)})`
-    }
-  },
-  thumbsSection: {
-    display: 'none',
-    '@media (min-height:500px)': {
-      flex: '0 1 auto',
-      flexWrap: 'nowrap',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      width: '100vw'
-    }
-  },
-  thumbsSlide: {
-    display: 'flex',
-    flexWrap: 'nowrap',
-    overflowX: 'hidden',
-    overflowY: 'hidden',
-    alignItems: 'center',
-    gap: theme.spacing(2),
-    paddingLeft: theme.spacing(2),
-    paddingRight: theme.spacing(2)
-  },
-  thumbContainer: {
-    color: 'white',
-    backgroundColor: '#00000080',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    cursor: 'pointer',
-    '@media (min-height:500px)': {
-      borderRadius: theme.spacing(1),
-      minHeight: MDSize,
-      minWidth: MDSize,
-      maxHeight: MDSize,
-      maxWidth: MDSize,
-      overflow: 'hidden'
-    },
-    '@media (min-height:720px)': {
-      borderRadius: theme.spacing(1.5),
-      minHeight: LGSize,
-      minWidth: LGSize,
-      maxHeight: LGSize,
-      maxWidth: LGSize,
-      overflow: 'hidden'
-    },
-    '@media (min-height:1080px)': {
-      borderRadius: theme.spacing(2),
-      minHeight: XLSize,
-      minWidth: XLSize,
-      maxHeight: XLSize,
-      maxWidth: XLSize,
-      overflow: 'hidden'
-    }
-  },
-  thumb: {
-    display: 'flex',
-    '@media (min-height:500px)': {
-      maxHeight: MDSize
-    },
-    '@media (min-height:720px)': {
-      maxHeight: LGSize
-    },
-    '@media (min-height:1080px)': {
-      maxHeight: XLSize
-    }
-  },
-  unselectedThumb: {
-    filter: `brightness(50%)`,
-    transition: theme.transitions.create('filter', {
-      easing: theme.transitions.easing.easeInOut,
-      duration: theme.transitions.duration.short
-    }),
-    '&:hover': {
-      filter: `brightness(90%)`
-    }
-  },
-  button: {
-    '&:hover': {
-      backgroundColor: alpha(theme.palette.secondary.main, 0.5)
-    },
-    cursor: 'pointer',
-    backgroundColor: alpha(theme.palette.secondary.main, 0.3),
-    color: 'white',
-    margin: theme.spacing(2)
-  },
-  closeButton: {
-    position: 'absolute',
-    top: '0px',
-    right: '0px'
-  },
-  beforeButton: {
-    position: 'absolute',
-    top: '50%',
-    left: '0px',
-    transform: 'translate(0%, -50%)'
-  },
-  nextButton: {
-    position: 'absolute',
-    top: '50%',
-    right: '0px',
-    transform: 'translate(0%, -50%)'
-  },
-  imageLoading: {
-    color: 'white',
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderRadius: theme.spacing(1)
-  },
-  imageMissing: {
-    backgroundColor: '#00000080',
-    padding: theme.spacing(4)
-  }
-}));
-
-export type CarouselContextProps = {
-  carousel: boolean;
-  index: number;
-  images: Array<{
-    name: string;
-    description: string;
-    imgSrc: string;
-    thumbSrc: string;
-  }>;
-  setCarousel: (value: boolean) => void;
-  openCarousel: (
-    idx: number,
-    imgs: Array<{
-      name: string;
-      description: string;
-      imgSrc: string;
-      thumbSrc: string;
-    }>
-  ) => void;
-  closeCarousel: () => void;
-  setIndex: (value: number) => void;
-  setImages: (
-    images: Array<{
-      name: string;
-      description: string;
-      imgSrc: string;
-      thumbSrc: string;
-    }>
-  ) => void;
+export type Image = {
+  id?: string; // unique id using the simple hash
+  name: string;
+  description: string;
+  img: string; // sha256 of the image
+  thumb: string; // sha256 of the thumb
 };
+
+export const CAROUSEL_PARAM = 'carousel';
 
 export interface CarouselProviderProps {
   children: React.ReactNode;
 }
 
+export type CarouselContextProps = {
+  open?: boolean;
+  images: Image[];
+  onOpenImage: (id: string) => void;
+  onCloseImage: () => void;
+  onAddImage: (img: Image, onIDChange: (id: string) => void) => void;
+  onRemoveImage: (img: Image) => void;
+  onClearImages: () => void;
+  onNextImage?: () => void;
+  onPreviousImage?: () => void;
+};
+
 export const CarouselContext = React.createContext<CarouselContextProps>(null);
 
-function CarouselProvider(props: CarouselProviderProps) {
-  const classes = useStyles();
-  const { children } = props;
-  const [carousel, setCarousel] = useState<boolean>(false);
-  const [index, setIndex] = useState<number>(0);
-  const [images, setImages] = useState<
-    Array<{
-      name: string;
-      description: string;
-      imgSrc: string;
-      thumbSrc: string;
-    }>
-  >([]);
-  const carouselItemsRef = useRef<HTMLDivElement[] | null[]>([]);
+const WrappedCarouselProvider = ({ children }: CarouselProviderProps) => {
+  const location = useLocation();
+  const navigate = useNavigate();
+  const { user: currentUser } = useALContext();
 
-  const openCarousel = (idx, imgs) => {
-    setIndex(idx);
-    setImages(imgs);
-    setCarousel(true);
-  };
+  const [images, setImages] = useState<Image[]>([]);
+  const [open, setOpen] = useState<boolean>(false);
 
-  const closeCarousel = () => {
-    setCarousel(false);
-    setIndex(0);
-    setImages([]);
-  };
+  const simpleHash = useCallback((text: string) => {
+    let hash = 0;
+    for (var i = 0; i < text.length; i++) {
+      var char = text.charCodeAt(i);
+      hash = (hash << 5) - hash + char;
+      hash = hash & hash; // Convert to 32bit integer
+    }
+    return hash.toString(16);
+  }, []);
 
-  const handleSelectedImageChange = useCallback(
-    (newIndex: number) => {
-      if (images && images.length > 0) {
-        setIndex(newIndex);
-        if (carouselItemsRef?.current[newIndex]) {
-          carouselItemsRef?.current[newIndex]?.scrollIntoView({
-            inline: 'center',
-            behavior: 'smooth'
-          });
-        }
-      }
+  const onOpenImage = useCallback(
+    (id: string) => {
+      const query = new SimpleSearchQuery(location.search);
+      query.set(CAROUSEL_PARAM, id);
+      navigate(`${location.pathname}?${query.toString()}${location.hash}`);
     },
-    [images]
+    [location.hash, location.pathname, location.search, navigate]
   );
 
-  const onPreviousImage = useCallback(() => {
-    if (images && images.length > 0) {
-      let newIndex = index >= 1 ? index - 1 : images.length - 1;
-      handleSelectedImageChange(newIndex);
-    }
-  }, [handleSelectedImageChange, images, index]);
+  const onCloseImage = useCallback(() => {
+    const query = new SimpleSearchQuery(location.search);
+    query.pop(CAROUSEL_PARAM);
+    navigate(`${location.pathname}?${query.toString()}${location.hash}`);
+  }, [location.hash, location.pathname, location.search, navigate]);
+
+  const onAddImage = useCallback(
+    (img: Image, onIDChange: (id: string) => void) => {
+      setImages(imgs => {
+        if (imgs.some(i => i.name === img.name && i.img === img.img)) return imgs;
+        const count = imgs.filter(
+          i => i.name === img.name || i.description === img.description || i.img === img.img
+        ).length;
+        const id = simpleHash(`${JSON.stringify(img)}${count}`);
+        onIDChange(id);
+        return [...imgs, { ...img, id }];
+      });
+    },
+    [simpleHash]
+  );
+
+  const onRemoveImage = useCallback(
+    (img: Image) => setImages(imgs => imgs.filter(i => i.name !== img.name || i.img !== img.img)),
+    []
+  );
+
+  const onClearImages = useCallback(() => setImages([]), []);
 
   const onNextImage = useCallback(() => {
-    if (images && images.length > 0) {
-      let newIndex = index < images.length - 1 ? index + 1 : 0;
-      handleSelectedImageChange(newIndex);
-    }
-  }, [handleSelectedImageChange, images, index]);
+    const query = new SimpleSearchQuery(location.search);
+    const param = query.get(CAROUSEL_PARAM, null);
+    const index = images.findIndex(i => i.id === param);
 
-  useEffect(() => {
-    if (images && images[index]) {
-      carouselItemsRef.current = carouselItemsRef.current.slice(0, images.length);
-      handleSelectedImageChange(index);
+    if (index >= 0 && index < images.length - 1) {
+      query.set(CAROUSEL_PARAM, images[index + 1].id);
+      navigate(`${location.pathname}?${query.toString()}${location.hash}`);
     }
-  }, [index, images, handleSelectedImageChange]);
+  }, [images, location.hash, location.pathname, location.search, navigate]);
+
+  const onPreviousImage = useCallback(() => {
+    const query = new SimpleSearchQuery(location.search);
+    const param = query.get(CAROUSEL_PARAM, null);
+    const index = images.findIndex(i => i.id === param);
+
+    if (index >= 1 && index < images.length) {
+      query.set(CAROUSEL_PARAM, images[index - 1].id);
+      navigate(`${location.pathname}?${query.toString()}${location.hash}`);
+    }
+  }, [images, location.hash, location.pathname, location.search, navigate]);
+
+  useEffect(
+    () =>
+      setOpen(() => {
+        if (!currentUser) return false;
+        const query = new SimpleSearchQuery(location.search).get(CAROUSEL_PARAM, null);
+        return images.some(image => image.id === query);
+      }),
+    [currentUser, images, location.search]
+  );
 
   return (
     <CarouselContext.Provider
       value={{
-        carousel,
-        index,
+        open,
         images,
-        setCarousel,
-        openCarousel,
-        closeCarousel,
-        setIndex,
-        setImages
+        onAddImage,
+        onClearImages,
+        onCloseImage,
+        onNextImage,
+        onOpenImage,
+        onPreviousImage,
+        onRemoveImage
       }}
     >
-      {useMemo(
-        () => (
-          <>{children}</>
-        ),
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        [children]
-      )}
-      {useMemo(
-        () =>
-          carousel &&
-          images.length && (
-            <Drawer open classes={{ paper: classes.backdrop }}>
-              <Carousel
-                autofocus
-                enableSwipe
-                escapeCallback={closeCarousel}
-                onNext={onNextImage}
-                onPrevious={onPreviousImage}
-                style={{ height: '100%' }}
-              >
-                <div id="carousel-image" className={classes.carousel} tabIndex={-1}>
-                  <div className={classes.spacer} />
-                  <div className={classes.textContainer} style={{ paddingBottom: '4px' }}>
-                    <Typography className={classes.text} variant="body2">
-                      {images[index].name}
-                    </Typography>
-                  </div>
-                  <Image alt={images[index].name} src={images[index].imgSrc} />
-                  <div className={classes.textContainer} style={{ paddingTop: '4px' }}>
-                    <Typography className={classes.text} variant="body2">
-                      {images[index].description}
-                    </Typography>
-                  </div>
-                  <div className={classes.spacer} />
-                  <div className={classes.thumbsSection}>
-                    <div className={classes.thumbsSlide}>
-                      {images.map((element, i) => {
-                        return (
-                          <Image
-                            key={'thumb-' + i}
-                            alt={element.name}
-                            src={element.thumbSrc}
-                            isThumb
-                            thumbSelected={index === i}
-                            onClick={() => {
-                              handleSelectedImageChange(i);
-                            }}
-                            onRef={el => (carouselItemsRef.current[i] = el)}
-                          />
-                        );
-                      })}
-                    </div>
-                  </div>
-                </div>
-
-                <IconButton className={clsx(classes.button, classes.closeButton)} onClick={closeCarousel} size="large">
-                  <CloseOutlinedIcon />
-                </IconButton>
-
-                <IconButton
-                  className={clsx(classes.button, classes.beforeButton)}
-                  onClick={onPreviousImage}
-                  size="large"
-                >
-                  <NavigateBeforeIcon />
-                </IconButton>
-
-                <IconButton className={clsx(classes.button, classes.nextButton)} onClick={onNextImage} size="large">
-                  <NavigateNextIcon />
-                </IconButton>
-              </Carousel>
-            </Drawer>
-          ),
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-        [carousel, images, index]
-      )}
+      {useMemo(() => children, [children])}
+      <CarouselContainer />
     </CarouselContext.Provider>
   );
-}
+};
 
+export const CarouselProvider = React.memo(WrappedCarouselProvider);
 export default CarouselProvider;
-
-type ImageProps = {
-  id?: string;
-  alt: string;
-  src: string;
-  isThumb?: boolean;
-  thumbSelected?: boolean;
-  onClick?: () => void;
-  onRef?: (el: HTMLDivElement) => void;
-};
-
-const Image = ({ id, alt, src, isThumb, thumbSelected, onClick, onRef }: ImageProps) => {
-  const [image, setImage] = useState<string>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-  const classes = useStyles();
-  const { apiCall } = useMyAPI();
-  const { t } = useTranslation(['carousel']);
-
-  useEffect(() => {
-    setLoading(true);
-    setImage(null);
-    apiCall({
-      url: `/api/v4/file/image/${src}/`,
-      allowCache: true,
-      onSuccess: api_data => {
-        setLoading(false);
-        setImage(api_data.api_response);
-      },
-      onFailure: api_data => {
-        setLoading(false);
-        setImage(null);
-      }
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [alt, src]);
-
-  return (
-    <Tooltip title={isThumb ? alt : ''} placement="top">
-      <Box
-        className={
-          isThumb ? clsx(classes.thumbContainer, thumbSelected ? '' : classes.unselectedThumb) : classes.imageContainer
-        }
-        onClick={onClick}
-      >
-        {image ? (
-          <img id={id} className={isThumb ? classes.thumb : classes.image} src={image} alt={alt} ref={onRef} />
-        ) : loading ? (
-          <div className={isThumb ? classes.thumb : classes.imageLoading} ref={onRef}>
-            <CircularProgress color="secondary" />
-          </div>
-        ) : (
-          <div ref={onRef}>
-            <Box
-              className={isThumb ? classes.thumb : clsx(classes.imageLoading, classes.imageMissing)}
-              onClick={onClick}
-            >
-              <BrokenImageOutlinedIcon style={{ fontSize: isThumb ? '2rem' : '6rem' }} />
-              {!isThumb && <Typography>{t('not_found')}</Typography>}
-            </Box>
-          </div>
-        )}
-      </Box>
-    </Tooltip>
-  );
-};
