--- src/components/visual/Carousel/index.tsx
@@ -1,5 +1,0 @@
-export * from './Body';
-export * from './Container';
-export * from './Hook';
-export * from './Provider';
-export * from './Thumb';
