+++ src/components/visual/ResultCard/result_section.tsx
@@ -17,8 +17,8 @@ import Tag from 'components/visual/Tag';
 import Verdict from 'components/visual/Verdict';
 import React, { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ImageBody } from '../Carousel';
 import { GraphBody } from './graph_body';
+import { ImageBody } from './image_body';
 import { InvalidBody } from './invalid_body';
 import { JSONBody } from './json_body';
 import { KVBody } from './kv_body';
