--- src/locales/en/carousel.json
@@ -1,3 +1,0 @@
-{
-  "not_found": "The API is preventing us from displaying this image."
-}
