--- src/locales/fr/carousel.json
@@ -1,3 +1,0 @@
-{
-  "not_found": "L'API nous empêche d'afficher cette image."
-}
